+++ src/main/java/com/plutus/system/repository/ClientRepository.java
@@ -4,6 +4,8 @@ import com.plutus.system.model.entity.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, BigInteger> {
 }
